Modify player switch behavior.
This commit changes the player switching so that the controlled player
is always the player nearest to the ball also on restarts (except when
the restarter is the goalkeeper). It also adds a switch timer to remove
player switch blinking between two players.

This commit fixes the bug where there would be no restart on kickoff.
This bug was caused by the controlled player being outside the pitch,
and the player couldn't be moved due to restart.